refactor: Apply tailrec fun
