+++ src/main/kotlin/racingcar/domain/Car.kt
@@ -1,5 +1,7 @@
 package racingcar.domain
 
+import racingcar.vo.Position
+
 class Car(private var position: Position = Position()) {
 
     fun moveForward(moveStrategy: MoveStrategy): Position {
