+++ src/test/kotlin/racingcar/vo/RoundTest.kt
@@ -1,4 +1,4 @@
-package racingcar.domain
+package racingcar.vo
 
 import io.kotest.assertions.throwables.shouldNotThrowAny
 import io.kotest.assertions.throwables.shouldThrow
